Reduce CRM polling limit to make the counters available for route perf test
