+++ tests/route/test_route_perf.py
@@ -3,7 +3,10 @@ import json
 import logging
 from datetime import datetime
 from tests.common.utilities import wait_until
-from tests.common import config_reload
+from tests.common.helpers.assertions import pytest_assert
+
+CRM_POLL_INTERVAL = 1
+CRM_DEFAULT_POLL_INTERVAL = 300
 
 pytestmark = [
     pytest.mark.topology('any'),
@@ -13,42 +16,6 @@ pytestmark = [
 logger = logging.getLogger(__name__)
 
 ROUTE_TABLE_NAME = 'ASIC_STATE:SAI_OBJECT_TYPE_ROUTE_ENTRY'
-
-@pytest.fixture(autouse=True)
-def ignore_expected_loganalyzer_exceptions(duthost, loganalyzer):
-    """
-        Ignore expected failures logs during test execution.
-
-        The route_checker script will compare routes in APP_DB and ASIC_DB, and an ERROR will be
-        recorded if mismatch. The testcase will add 10,000 routes to APP_DB, and route_checker may
-        detect mismatch during this period. So a new pattern is added to ignore possible error logs.
-
-        Args:
-            duthost: DUT fixture
-            loganalyzer: Loganalyzer utility fixture
-    """
-    ignoreRegex = [
-        ".*ERR route_check.py:.*",
-        ".*ERR.* \'routeCheck\' status failed.*"
-    ]
-    if loganalyzer:
-        # Skip if loganalyzer is disabled
-        loganalyzer.ignore_regex.extend(ignoreRegex)
-
-@pytest.fixture(params=[4, 6])
-def ip_versions(request):
-    """
-    Parameterized fixture for IP versions.
-    """
-    yield request.param
-
-@pytest.fixture(scope='function', autouse=True)
-def reload_dut(duthost, request):
-    yield
-    if request.node.rep_call.failed:
-        #Issue a config_reload to clear statically added route table and ip addr
-        logging.info("Reloading config..")
-        config_reload(duthost)
 
 def prepare_dut(duthost, intf_neighs):
     for intf_neigh in intf_neighs:
@@ -179,9 +146,13 @@ def test_perf_add_remove_routes(duthost, request, ip_versions):
     intf_neighs, str_intf_nexthop = generate_intf_neigh(NUM_NEIGHS, ip_versions)
 
     route_tag = "ipv{}_route".format(ip_versions)
-    used_routes_count = duthost.get_crm_resources().get("main_resources").get(route_tag).get("used")
-    avail_routes_count = duthost.get_crm_resources().get("main_resources").get(route_tag).get("available")
+    used_routes_count = duthost.get_crm_resources().get("main_resources").get(route_tag, {}).get("used")
+    avail_routes_count = duthost.get_crm_resources().get("main_resources").get(route_tag, {}).get("available")
+    pytest_assert(avail_routes_count, "CRM main_resources data is not ready within adjusted CRM polling time {}s".\
+            format(CRM_POLL_INTERVAL))
+
     num_routes = min(avail_routes_count, set_num_routes)
+
     logger.info("IP route utilization before test start: Used: {}, Available: {}, Test count: {}"\
         .format(used_routes_count, avail_routes_count, num_routes))
